+++ src/main/java/co/edu/uniquindio/proyecto/proyectotienda/servicios/interfaces/VisitaServicio.java
@@ -2,5 +2,5 @@ package co.edu.uniquindio.proyecto.proyectotienda.servicios.interfaces;
 
 public interface VisitaServicio {
 
-    int cantidadVisitas(int codigoPublicacion) throws Exception;
+    Long cantidadVisitas(int codigoPublicacion) throws Exception;
 }
